split up models for algorithms and config
